+++ src/app/pass_checker/PasswordChecker.ts
@@ -3,6 +3,12 @@ export class PasswordChecker {
         if (password.length < 8) {
             return false;
         }
+        if (password === password.toLowerCase()) {
+            return false;
+        }
+        if (password === password.toUpperCase()) {
+            return false;
+        }
         return true;
     }
 }