+++ src/test/pass_checker/PasswordChecker.spec.ts
@@ -12,8 +12,29 @@ describe("PasswordChecker test suite", () => {
         expect(actual).toBe(false);
     });
 
+    // tests need change (not very ideal) to accommodate new requirements
     it("Password with more than 8 chars is ok", () => {
-        const actual = sut.checkPassword("12345678");
+        const actual = sut.checkPassword("12345678Aa");
+        expect(actual).toBe(true);
+    });
+
+    it("Password with no uppercase is invalid", () => {
+        const actual = sut.checkPassword("1234abcd");
+        expect(actual).toBe(false);
+    });
+
+    it("Password with uppercase is valid", () => {
+        const actual = sut.checkPassword("1234abcD");
+        expect(actual).toBe(true);
+    });
+
+    it("Password with no lowercase is invalid", () => {
+        const actual = sut.checkPassword("1234ABCD");
+        expect(actual).toBe(false);
+    });
+
+    it("Password with lowercase is valid", () => {
+        const actual = sut.checkPassword("1234ABCd");
         expect(actual).toBe(true);
     });
 });