+++ CHANGE_LOG.md
@@ -5,3 +5,7 @@
 -   length is less than 8 chars
 -   has no upper case letter
 -   has no lower case letter
+
+## Requirement 2:
+
+-   return the reasons that make a password invalid